Add a pass to try to remove empty words due to substitution.
The use of -S/-E/-I can lead to empty words, they must be deleted in all
cases except in column mode. In column mode, substitutions involving
empty words can lead to misalignments, so it is necessary to prohibit
them and terminate the program. These substitutions have to be made
with other tools before using this utility.