Add mouse support
- Direct selection of words by clicking.
- Tagging/untagging of words by clicking.
- Scrolling by clicks and wheel.
- Validation of the selection with a double-click.
- Shifting the content of the windows to the left or to the right,
  also added new keyboard keys < and > for that.

Three mouse tracking protocols are supported:
X11 (1000) and X11 SGR (1006) and urxvt (1015).

The X11 protocol limits lines and columns to 223 (e.g. screen < 4.7.0)
The other supported protocols don't have this limitation.

New options to disable the mouse, change the double-click delay and remap
the buttons of the mouse are available.

On (virtualised?) (Free)BSD it may be necessary to disable buttons
8 and 9, refer to the manual for details.